@@ -1,11 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 33
     defaultConfig {
         applicationId "me.pqpo.smartcropper"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 33
         versionCode 2
         versionName "1.1"
         ndk {
@@ -30,8 +30,8 @@ android {
 dependencies {
     implementation project(':smartcropperlib')
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     // For developers using AndroidX in their applications
     //implementation 'pub.devrel:easypermissions:3.0.0'
 
